[ext-msg] split into MessagingCommander and NotificationCommander
