+++ kifi-backend/modules/eliza/app/com/keepit/eliza/controllers/site/WebsiteMessagingController.scala
@@ -1,6 +1,6 @@
 package com.keepit.eliza.controllers.site
 
-import com.keepit.eliza.commanders.MessagingCommander
+import com.keepit.eliza.commanders.{NotificationCommander, MessagingCommander}
 import com.keepit.common.controller.{WebsiteController, ElizaServiceController, ActionAuthenticator}
 import com.keepit.common.time._
 import com.keepit.heimdal._
@@ -13,6 +13,7 @@ import com.google.inject.Inject
 
 class WebsiteMessagingController @Inject() (
   messagingCommander: MessagingCommander,
+  notificationCommander: NotificationCommander,
   actionAuthenticator: ActionAuthenticator,
   heimdalContextBuilder: HeimdalContextBuilderFactory
   ) extends WebsiteController(actionAuthenticator) with ElizaServiceController {
@@ -20,9 +21,9 @@ class WebsiteMessagingController @Inject() (
   def getNotifications(howMany: Int, before: Option[String]) = JsonAction.authenticatedAsync { request =>
     val noticesFuture = before match {
       case Some(before) =>
-        messagingCommander.getSendableNotificationsBefore(request.userId, parseStandardTime(before), howMany.toInt)
+        notificationCommander.getSendableNotificationsBefore(request.userId, parseStandardTime(before), howMany.toInt)
       case None =>
-        messagingCommander.getLatestSendableNotifications(request.userId, howMany.toInt)
+        notificationCommander.getLatestSendableNotifications(request.userId, howMany.toInt)
     }
     noticesFuture.map {notices =>
       val numUnreadUnmuted = messagingCommander.getUnreadUnmutedThreadCount(request.userId)
