+++ kifi-backend/modules/eliza/test/com/keepit/eliza/MessagingTest.scala
@@ -20,7 +20,7 @@ import com.keepit.common.healthcheck.FakeAirbrakeNotifier
 import com.keepit.abook.{FakeABookServiceClientImpl, ABookServiceClient, TestABookServiceClientModule}
 
 import com.keepit.eliza.controllers.WebSocketRouter
-import com.keepit.eliza.commanders.MessagingCommander
+import com.keepit.eliza.commanders.{NotificationCommander, MessagingCommander}
 import com.keepit.eliza.controllers.internal.MessagingController
 import com.keepit.eliza.model._
 
@@ -84,11 +84,12 @@ class MessagingTest extends Specification with DbTestInjector {
 
         val (user1, user2, user3, user2n3Seq, shoebox) = setup()
         val messagingCommander = inject[MessagingCommander]
+        val notificationCommander = inject[NotificationCommander]
         val (thread1, msg1) = messagingCommander.sendNewMessage(user1, user2n3Seq, Nil, Json.obj("url" -> "http://thenextgoogle.com"), Some("title"), "World!")
 
         val (thread2, msg2) = messagingCommander.sendMessage(user1, msg1.thread, "Domination!", None)
 
-        Await.result(messagingCommander.getLatestSendableNotifications(user1, 20), Duration(4, "seconds")).jsons.length === 1
+        Await.result(notificationCommander.getLatestSendableNotifications(user1, 20), Duration(4, "seconds")).jsons.length === 1
 
         val messageIds : Seq[Option[Id[Message]]] = messagingCommander.getThreads(user2).flatMap(messagingCommander.getThreadMessages(_, None)).map(_.id)
         val messageContents : Seq[String] = messagingCommander.getThreads(user2).flatMap(messagingCommander.getThreadMessages(_, None)).map(_.messageText)
@@ -108,6 +109,7 @@ class MessagingTest extends Specification with DbTestInjector {
 
         val (user1, user2, user3, user2n3Seq, shoebox) = setup()
         val messagingCommander = inject[MessagingCommander]
+        val notificationCommander = inject[NotificationCommander]
         var notified = scala.collection.concurrent.TrieMap[Id[User], Int]()
 
         inject[WebSocketRouter].onNotification{ (userId, notification) =>
@@ -127,12 +129,12 @@ class MessagingTest extends Specification with DbTestInjector {
         notified.isDefinedAt(user1) === false
         notified(user2) === 2
 
-        messagingCommander.getLatestSendableNotifications(user3, 10)
+        notificationCommander.getLatestSendableNotifications(user3, 10)
 
-        messagingCommander.getUnreadThreadNotifications(user3).length === 1 //there was only one thread created due to merging
+        notificationCommander.getUnreadThreadNotifications(user3).length === 1 //there was only one thread created due to merging
         messagingCommander.getUnreadUnmutedThreadCount(user3) === 1
 
-        val notifications : Seq[JsObject] = Await.result(messagingCommander.getLatestUnreadSendableNotifications(user3, 20), Duration(4, "seconds"))._1.jsons
+        val notifications : Seq[JsObject] = Await.result(notificationCommander.getLatestUnreadSendableNotifications(user3, 20), Duration(4, "seconds"))._1.jsons
         notifications.length === 1
         val participants = (notifications.head \ "participants").as[Seq[BasicUser]].sortBy (_.lastName)
         println(participants)
@@ -141,8 +143,8 @@ class MessagingTest extends Specification with DbTestInjector {
         participants(1).lastName.endsWith(user2.id.toString) === true
         participants(2).lastName.endsWith(user3.id.toString) === true
 
-        messagingCommander.setAllNotificationsRead(user3)
-        messagingCommander.getUnreadThreadNotifications(user3).length === 0
+        notificationCommander.setAllNotificationsRead(user3)
+        notificationCommander.getUnreadThreadNotifications(user3).length === 0
         messagingCommander.getUnreadUnmutedThreadCount(user3) === 0
 
 
@@ -154,17 +156,18 @@ class MessagingTest extends Specification with DbTestInjector {
 
         val (user1, user2, user3, user2n3Seq, shoebox) = setup()
         val messagingCommander = inject[MessagingCommander]
+        val notificationCommander = inject[NotificationCommander]
 
         val (thread, msg) = messagingCommander.sendNewMessage(user1, Seq(user2), Nil, Json.obj("url" -> "http://kifi.com"), Some("title"), "Fortytwo")
 
         Thread.sleep(100) //AHHHHHH. Really need to figure out how to test Async code with multiple execution contexts. (https://app.asana.com/0/5674704693855/9223435240746)
-        Await.result(messagingCommander.getLatestSendableNotifications(user2, 1), Duration(4, "seconds")).jsons.length === 1
-        Await.result(messagingCommander.getLatestSendableNotifications(user3, 1), Duration(4, "seconds")).jsons.length === 0
+        Await.result(notificationCommander.getLatestSendableNotifications(user2, 1), Duration(4, "seconds")).jsons.length === 1
+        Await.result(notificationCommander.getLatestSendableNotifications(user3, 1), Duration(4, "seconds")).jsons.length === 0
 
         val user3ExtId = Await.result(shoebox.getUser(user3), Duration(4, "seconds")).get.externalId
         messagingCommander.addParticipantsToThread(user1, thread.externalId, Seq(user3ExtId))
         Thread.sleep(200) //See comment for same above
-        Await.result(messagingCommander.getLatestSendableNotifications(user3, 1), Duration(4, "seconds")).jsons.length === 1
+        Await.result(notificationCommander.getLatestSendableNotifications(user3, 1), Duration(4, "seconds")).jsons.length === 1
       }
     }
